+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added:
+- Added `attachmentSizeLimit` to interactions
 ## [6.3.2] - 2025-03-29
 ### Fixed:
 - Allow for just `components` in `MessageInteractionContext.send` and `MessageInteractionContext.sendFollowUp`